# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.SkillRequest:
    properties:
      category:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
  dto.SkillResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UserSkillRequest:
    properties:
      level:
        type: integer
      skill_id:
        type: integer
    type: object
  dto.UserSkillResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      level:
        type: integer
      skill_id:
        type: integer
      skill_name:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API для системы управления проектами и задачами с поддержкой навыков
    сотрудников
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SkillsTracker API
  version: 1.0.0
paths:
...
  /api/v1/skills:
    get:
      description: Получить список всех навыков или навыков по категории. Если указана
        категория, возвращаются только навыки этой категории.
      parameters:
      - description: 'Категория навыка (например: technical, soft, domain)'
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SkillResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить навыки
      tags:
      - skills
    post:
      consumes:
      - application/json
      description: Создает новый навык в системе. Навыки используются для сопоставления
        с задачами и пользователями.
      parameters:
      - description: Данные навыка (name, description, category)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SkillRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SkillResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать навык
      tags:
      - skills
  /api/v1/skills/{id}:
    delete:
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить навык
      tags:
      - skills
    get:
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SkillResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить навык по id
      tags:
      - skills
    put:
      consumes:
      - application/json
      parameters:
      - description: Skill ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные навыка
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.SkillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить навык
      tags:
      - skills
...
  /api/v1/users/{user_id}/skills:
    get:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserSkillResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить навыки пользователя
      tags:
      - skills
    post:
      consumes:
      - application/json
      description: Добавляет навык пользователю с указанным уровнем (1-5). Если навык
        уже есть у пользователя, обновляется уровень.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Данные навыка (skill_id, level от 1 до 5)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserSkillRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserSkillResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить навык пользователю
      tags:
      - skills
  /api/v1/users/{user_id}/skills/{skill_id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить навык у пользователя
      tags:
      - skills
    put:
      consumes:
      - application/json
      description: Обновляет уровень навыка пользователя (1-5). Навык должен уже быть
        добавлен пользователю.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Skill ID
        in: path
        name: skill_id
        required: true
        type: integer
      - description: Уровень навыка (1-5)
        in: query
        name: level
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить уровень навыка пользователя
      tags:
      - skills
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"